backend/requirements.txt
flask>=2.0.1
flask-sqlalchemy>=2.5.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
openai>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
python-jose>=3.3.0
passlib>=1.7.4
mysqlclient>=2.1.0
mysql-connector-python>=8.0.26